% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccc.R
\name{ccc}
\alias{ccc}
\title{An Rcpp implementation of Lin's Concordance Correlation Coefficient (CCC)}
\usage{
ccc(x, y)
}
\arguments{
\item{x}{A numeric vector containing \code{x}.}

\item{y}{A numeric vector containing \code{y}.}
}
\value{
A numeric scalar representing the CCC between \code{x} and \code{y}.
}
\description{
An Rcpp implementation of Lin's Concordance Correlation Coefficient (CCC)
}
\examples{
# Create x and y vectors
x <- rnorm(100, sd = 1)
y <- rnorm(length(x), sd = 0.2)

# Calculate CCC between x and y
ccc(x,y)

}
